+++ CHANGELOG.md
@@ -1,3 +1,66 @@
+7.0.0 / 2017-01-13
+==================
+  - Update docgen.js
+  - Upgrade packages in package.json
+  - Add yarn.lock file
+  - Remove standalone version
+  - Update README.md
+  - Add `array-callback-return`
+  - Add `class-method-use-this`
+  - Add `no-case-declarations`
+  - Add `no-empty-function`
+  - Add `no-global-assign`
+  - Add `no-magic-numbers`
+  - Remove `no-process-exit` from rules/best-practices.js
+  - Add `no-alert`
+  - Add `no-extra-label`
+  - Add `no-return-await`
+  - Add `no-self-assign`
+  - Add `no-unmodified-loop-condition`
+  - Add `no-unused-labels`
+  - Add `no-useless-concat`
+  - Add `no-useless-escape`
+  - Add `no-useless-return`
+  - Add `require-await`
+  - Add `no-restricted-properties`
+  - Add `no-restricted-globals`
+  - Add `global-require`
+  - Add `no-process-env`
+  - Add `no-await-in-loop`
+  - Add `no-prototype-builtins`
+  - Add `no-template-curly-in-string`
+  - Add `no-unsafe-finally`
+  - Add `no-unsafe-negation`
+  - Add `no-useless-rename`
+  - Add `prefer-destructuring`
+  - Add `prefer-numeric-literals`
+  - Add `rest-spread-spacing`
+  - Add `symbol-description`
+  - Add `block-spacing`
+  - Add `capitalized-comments`
+  - Add `func-call-spacing`
+  - Add `func-name-matching`
+  - Add `id-blacklist`
+  - Add `jsx-quotes`
+  - Add `line-comment-position`
+  - Add `lines-around-directive`
+  - Add `max-lines`
+  - Add `max-statements-per-line`
+  - Add `multiline-ternary`
+  - Add `newline-before-return`
+  - Add `newline-per-chained-call`
+  - Add `no-mixed-operators`
+  - Add `no-negated-condition`
+  - Add `no-restricted-syntax`
+  - Add `no-tabs`
+  - Add `no-whitespace-before-property`
+  - Add `object-curly-newline`
+  - Add `object-property-newline`
+  - Add `one-var-declaration-per-line`
+  - Add `require-jsdoc`
+  - Add `sort-keys`
+  - Add `unicode-bom`
+
 6.0.0 / 2016-07-27
 ==================
   - Create .npmignore